requirements.txt
python-dotenv
setuptools
tinydb
discord.py
attrs
cattrs